reorganize
- move vcc/core/utils to vcc/utils
- move vcc/core/cartridges to vcc/cartridges
- move cartridge_loader from core to utils
- move rom_image from core to vcc/devices/rom
- move banked_rom_image from mpi pak to vcc/devices/rom
