+++ libcommon/include/vcc/utils/configuration_serializer.h
@@ -16,45 +16,33 @@
 //	VCC (Virtual Color Computer). If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
 #pragma once
-#include <vcc/core/cartridge.h>
+#include <vcc/core/detail/exports.h>
+#include <string>
 
 
-namespace vcc { namespace core { namespace cartridges
+namespace vcc::utils
 {
 
-	struct LIBCOMMON_EXPORT basic_cartridge : public ::vcc::core::cartridge
+	class configuration_serializer
 	{
 	public:
 
-		using name_type = std::string;
-		using catalog_id_type = std::string;
-
+		using path_type = ::std::string;
+		using string_type = ::std::string;
 
 	public:
 
-		using cartridge::cartridge;
+		LIBCOMMON_EXPORT explicit configuration_serializer(path_type path);
 
-		name_type name() const override;
-		catalog_id_type catalog_id() const override;
-		description_type description() const override;
+		LIBCOMMON_EXPORT void write(const string_type& section, const string_type& key, int value) const;
+		LIBCOMMON_EXPORT void write(const string_type& section, const string_type& key, const string_type& value) const;
 
-		void start() override;
-		void stop() override;
-
-		void reset() override;
-		void process_horizontal_sync() override;
-		void write_port(unsigned char port_id, unsigned char value) override;
-		unsigned char read_port(unsigned char port_id) override;
-		unsigned char read_memory_byte(unsigned short memory_address) override;
-		void status(char* text_buffer, size_t buffer_size) override;
-		unsigned short sample_audio() override;
-		void menu_item_clicked(unsigned char menu_item_id) override;
+		LIBCOMMON_EXPORT int read(const string_type& section, const string_type& key, int default_value) const;
+		LIBCOMMON_EXPORT string_type read(const string_type& section, const string_type& key, const string_type& default_value = {}) const;
 
 
-	protected:
+	private:
 
-		virtual void initialize_pak();
-		virtual void initialize_bus();
+		const path_type path_;
 	};
-
-} } }
+}
