+++ libcommon/include/vcc/utils/filesystem.h
@@ -15,40 +15,19 @@
 //	You should have received a copy of the GNU General Public License along with
 //	VCC (Virtual Color Computer). If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
-#include <vcc/core/rom_image.h>
-#include <fstream>
+#pragma once
+#include <vcc/core/detail/exports.h>
+#include <string>
 #include <Windows.h>
 
 
-namespace vcc { namespace core
+namespace vcc::utils
 {
 
-	bool rom_image::load(path_type filename)
-	{
-		std::ifstream input(filename, std::ios::binary);
-		if (!input.is_open())
-		{
-			return false;
-		}
+	LIBCOMMON_EXPORT std::string get_module_path(HMODULE module_handle = nullptr);
+	LIBCOMMON_EXPORT std::string find_pak_module_path(std::string path);
+	LIBCOMMON_EXPORT std::string get_directory_from_path(std::string path);
+	LIBCOMMON_EXPORT std::string get_filename(std::string path);
+	LIBCOMMON_EXPORT std::string strip_application_path(std::string path);
 
-		std::streamoff begin = input.tellg();
-		input.seekg(0, std::ios::end);
-		std::streamoff end = input.tellg();
-		std::streamoff file_length = end - begin;
-		input.seekg(0, std::ios::beg);
-
-		std::vector<unsigned char> rom_data(static_cast<size_t>(file_length));
-		input.read(reinterpret_cast<char*>(&rom_data[0]), rom_data.size());
-
-		if (file_length != rom_data.size())
-		{
-			return false;
-		}
-
-		filename_ = move(filename);
-		data_ = move(rom_data);
-
-		return true;
-	}
-
-} }
+}
