+++ libcommon/libcommon.vcxproj
@@ -98,47 +98,48 @@
     <Link />
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="src\core\banked_rom_image.cpp" />
-    <ClCompile Include="src\core\cartridges\basic_cartridge.cpp" />
-    <ClCompile Include="src\core\cartridges\legacy_cartridge.cpp" />
-    <ClCompile Include="src\core\cartridges\rom_cartridge.cpp" />
-    <ClCompile Include="src\core\cartridge_loader.cpp" />
-    <ClCompile Include="src\core\rom_image.cpp" />
-    <ClCompile Include="src\core\utils\configuration_serializer.cpp" />
-    <ClCompile Include="src\core\utils\filesystem.cpp" />
-    <ClCompile Include="src\core\utils\winapi.cpp" />
+    <ClCompile Include="src\cartridges\basic_cartridge.cpp" />
+    <ClCompile Include="src\cartridges\legacy_cartridge.cpp" />
+    <ClCompile Include="src\cartridges\rom_cartridge.cpp" />
+    <ClCompile Include="src\devices\rom\banked_rom_image.cpp" />
+    <ClCompile Include="src\devices\rom\rom_image.cpp" />
     <ClCompile Include="src\devices\rtc\cloud9.cpp" />
     <ClCompile Include="src\devices\rtc\oki_m6242b.cpp" />
     <ClCompile Include="src\DialogOps.cpp" />
     <ClCompile Include="src\Fileops.cpp" />
     <ClCompile Include="src\logger.cpp" />
     <ClCompile Include="src\main.cpp" />
+    <ClCompile Include="src\utils\cartridge_loader.cpp" />
+    <ClCompile Include="src\utils\configuration_serializer.cpp" />
+    <ClCompile Include="src\utils\filesystem.cpp" />
+    <ClCompile Include="src\utils\winapi.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="include\vcc\cartridges\basic_cartridge.h" />
+    <ClInclude Include="include\vcc\cartridges\legacy_cartridge.h" />
+    <ClInclude Include="include\vcc\cartridges\null_cartridge.h" />
+    <ClInclude Include="include\vcc\cartridges\rom_cartridge.h" />
     <ClInclude Include="include\vcc\common\DialogOps.h" />
     <ClInclude Include="include\vcc\common\FileOps.h" />
     <ClInclude Include="include\vcc\common\logger.h" />
     <ClInclude Include="include\vcc\core\banked_rom_image.h" />
     <ClInclude Include="include\vcc\core\cartridge.h" />
-    <ClInclude Include="include\vcc\core\cartridges\basic_cartridge.h" />
-    <ClInclude Include="include\vcc\core\cartridges\legacy_cartridge.h" />
-    <ClInclude Include="include\vcc\core\cartridges\null_cartridge.h" />
-    <ClInclude Include="include\vcc\core\cartridges\rom_cartridge.h" />
     <ClInclude Include="include\vcc\core\cartridge_context.h" />
     <ClInclude Include="include\vcc\core\cartridge_factory.h" />
-    <ClInclude Include="include\vcc\core\cartridge_loader.h" />
     <ClInclude Include="include\vcc\core\detail\exports.h" />
     <ClInclude Include="include\vcc\core\interrupts.h" />
     <ClInclude Include="include\vcc\core\legacy_cartridge_definitions.h" />
     <ClInclude Include="include\vcc\core\limits.h" />
-    <ClInclude Include="include\vcc\core\rom_image.h" />
-    <ClInclude Include="include\vcc\core\utils\configuration_serializer.h" />
-    <ClInclude Include="include\vcc\core\utils\critical_section.h" />
-    <ClInclude Include="include\vcc\core\utils\dll_deleter.h" />
-    <ClInclude Include="include\vcc\core\utils\filesystem.h" />
-    <ClInclude Include="include\vcc\core\utils\winapi.h" />
+    <ClInclude Include="include\vcc\devices\rom\banked_rom_image.h" />
+    <ClInclude Include="include\vcc\devices\rom\rom_image.h" />
     <ClInclude Include="include\vcc\devices\rtc\cloud9.h" />
     <ClInclude Include="include\vcc\devices\rtc\oki_m6242b.h" />
+    <ClInclude Include="include\vcc\utils\cartridge_loader.h" />
+    <ClInclude Include="include\vcc\utils\configuration_serializer.h" />
+    <ClInclude Include="include\vcc\utils\critical_section.h" />
+    <ClInclude Include="include\vcc\utils\dll_deleter.h" />
+    <ClInclude Include="include\vcc\utils\filesystem.h" />
+    <ClInclude Include="include\vcc\utils\winapi.h" />
     <ClInclude Include="resource\resource.h" />
   </ItemGroup>
   <ItemGroup>
