+++ libcommon/src/cartridges/rom_cartridge.cpp
@@ -15,10 +15,10 @@
 //	You should have received a copy of the GNU General Public License along with
 //	VCC (Virtual Color Computer). If not, see <http://www.gnu.org/licenses/>.
 ////////////////////////////////////////////////////////////////////////////////
-#include <vcc/core/cartridges/rom_cartridge.h>
+#include <vcc/cartridges/rom_cartridge.h>
 
 
-namespace vcc { namespace core { namespace cartridges
+namespace vcc::cartridges
 {
 
 	rom_cartridge::rom_cartridge(
@@ -76,4 +76,4 @@ namespace vcc { namespace core { namespace cartridges
 		context_->assert_cartridge_line(true);
 	}
 
-} } }
+}
