+++ pakinterface.h
@@ -17,14 +17,14 @@ This file is part of VCC (Virtual Color Computer).
 */
 
 #pragma once
-#include <vcc/core/cartridge_loader.h>
+#include <vcc/utils/cartridge_loader.h>
 
 void PakTimer();
 unsigned char PakReadPort (unsigned char);
 void PakWritePort(unsigned char,unsigned char);
 unsigned char PackMem8Read (unsigned short);
 void GetModuleStatus( SystemState *);
-vcc::core::cartridge_loader_status PakLoadCartridge(const char* filename);
+vcc::utils::cartridge_loader_status PakLoadCartridge(const char* filename);
 void PakLoadCartridgeUI();
 unsigned short PackAudioSample();
 void ResetBus();
